backend/requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
openai-whisper
pydub